# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open Food Facts Parser API",
        "description": "API REST para curadoria de dados do Open Food Facts",
        "contact": {
            "email": "suporte@fitnessfood.lc"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Principal"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Health"
                ],
                "summary": "Retorna o status da API",
                "operationId": "6b71d5e5017cd63483d0142ee3ccea62",
                "responses": {
                    "200": {
                        "description": "Dados de saúde do sistema"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista produtos com paginação",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/products/{code}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtém um produto pelo código",
                "operationId": "49b03d87d85276e8020558b6ee931937",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualiza um produto",
                "operationId": "3ca6aab99570e78efd32bac1e792c4d2",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "trash",
                                            "published"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Muda o status do produto para 'trash'",
                "operationId": "1e3eeae3513271993540a6fa928cbc6f",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status atualizado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Health",
            "description": "Health"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}